// types/enums.ts
export enum LeetcodeDifficulty {
  EASY = 'easy',
  MEDIUM = 'medium',
  HARD = 'hard',
}

export enum PersonalDifficulty {
  TRIVIAL = 1, // Knew it instantly
  EASY = 2, // Minor hints needed
  MEDIUM = 3, // Significant thinking
  HARD = 4, // Struggled considerably
  IMPOSSIBLE = 5, // Could not solve
}